+++ CHANGELOG.md
@@ -1,47 +1,74 @@
 # Changelog
 
+## v1.4
+
+### v1.4.1 [![14.09.2022](https://img.shields.io/date/1663154237)](https://github.com/d8corp/watch-state-history-api/tree/v1.4.1)
+
+- change readme
+
+### v1.4.0 [![14.09.2022](https://img.shields.io/date/1663154096)](https://github.com/d8corp/watch-state-history-api/tree/v1.4.0)
+
+- `search` renames to `getSearch`
+- add `search` field
+
 ## v1.3
+
 ### v1.3.6 [![07.08.2021](https://img.shields.io/date/1628366228)](https://github.com/d8corp/watch-state-history-api/tree/v1.3.6)
+
 - updated dependencies
 
 ### v1.3.5 [![26.07.2021](https://img.shields.io/date/1627328144)](https://github.com/d8corp/watch-state-history-api/tree/v1.3.5)
+
 - hot fixes of the build flow
 
 ### v1.3.4 [![26.07.2021](https://img.shields.io/date/1627327647)](https://github.com/d8corp/watch-state-history-api/tree/v1.3.4)
+
 - updated [@watch-state/decorators](https://github.com/d8corp/watch-state-decorators/tree/v1.2.0) version to `v1.2`
 - fixed bag when you push the same url as you have and history state is empty (*first time opened page*)
 
 ### v1.3.3 [![10.07.2021](https://img.shields.io/date/1625947964)](https://github.com/d8corp/watch-state-history-api/tree/v1.3.3)
+
 - updated [watch-state](https://github.com/d8corp/watch-state/releases/tag/v3.3.3) version
 
 ### v1.3.2 [![27.06.2021](https://img.shields.io/date/1624823403)](https://github.com/d8corp/watch-state-history-api/tree/v1.3.2)
+
 - changed file structure
 - fixed types location
 
 ### v1.3.1 [![27.06.2021](https://img.shields.io/date/1624786974)](https://github.com/d8corp/watch-state-history-api/tree/v1.3.1)
+
 - fixed `tslib` dependency
 
 ### v1.3.0 [![27.06.2021](https://img.shields.io/date/1624785965)](https://github.com/d8corp/watch-state-history-api/tree/v1.3.0)
+
 - updated [watch-state](https://www.npmjs.com/package/watch-state) version 3
 - added current version by default key of `History`
 
 ## v1.2
+
 ### v1.2.1 [![13.02.2021](https://img.shields.io/date/1613234105)](https://github.com/d8corp/watch-state-history-api/tree/v1.2.1)
+
 - fixed a bug of history back action scroll
 
 ### v1.2.0 [![11.01.2021](https://img.shields.io/date/1610394742)](https://github.com/d8corp/watch-state-history-api/tree/v1.2.0)
+
 - updated `watch-state` to the second version
 
 ## v1.1
+
 ### v1.1.0 [![12.12.2020](https://img.shields.io/date/1607798652)](https://github.com/d8corp/watch-state-history-api/tree/v1.1.0)
+
 - now you can use number and boolean as value of query param in `setSearch`
 - now you can use an object as the second argument of `setSearch`
 - removed `scroll` from `History`
 - updated `watch-state` to `v1.1`
 
 ## v1.0
+
 ### v1.0.1 [![21.11.2020](https://img.shields.io/date/1605959328)](https://github.com/d8corp/watch-state-history-api/tree/v1.0.1)
+
 - updated dependencies and readme
 
 ### v1.0.0 [![19.11.2020](https://img.shields.io/date/1605817710)](https://github.com/d8corp/watch-state-history-api/tree/v1.0.0)
+
 - first implementation